# untitled/src/backend/openapi/starbooks-api.yaml
openapi: 3.0.3
info:
  title: StarBooks Backend API
  version: 1.0.0
  description: |
    StarBooks 독서 커뮤니티 백엔드 REST API 명세.
    - 인증: JWT (Access + Refresh)
    - 요청/응답 예시는 SDS, SRS, 클래스 다이어그램 요구사항을 반영함
servers:
  - url: https://api.starbooks.com
    description: Production
  - url: https://staging-api.starbooks.com
    description: Staging
...
paths:
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: 회원가입
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 회원 생성
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: 중복 아이디/닉네임
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 토큰 발급
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: 인증 실패
...
  /api/v1/books:
    get:
      tags: [Books]
      summary: 도서 검색
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookSummary'
    post:
      tags: [Books]
      security:
        - bearerAuth: []
      summary: 도서 등록 (관리자)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreateRequest'
      responses:
        '201':
          description: 생성
  /api/v1/books/{bookId}:
    parameters:
      - $ref: '#/components/parameters/BookId'
    get:
      tags: [Books]
      summary: 도서 상세
      responses:
        '200':
          description: 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDetail'
        '404':
          description: 없음
    patch:
      tags: [Books]
      security:
        - bearerAuth: []
      summary: 도서 정보 수정
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookUpdateRequest'
      responses:
        '200':
          description: 수정됨
    delete:
      tags: [Books]
      security:
        - bearerAuth: []
      summary: 도서 삭제
      responses:
        '204':
          description: 삭제
...
  /api/v1/challenges:
    get:
      tags: [Challenges]
      summary: 챌린지 목록
      responses:
        '200':
          description: 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChallengeSummary'
    post:
      tags: [Challenges]
      security:
        - bearerAuth: []
      summary: 챌린지 생성 (관리자)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeCreateRequest'
      responses:
        '201':
          description: 생성
  /api/v1/challenges/{challengeId}:
    parameters:
      - $ref: '#/components/parameters/ChallengeId'
    get:
      tags: [Challenges]
      summary: 챌린지 상세
      responses:
        '200':
          description: 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeDetail'
    patch:
      tags: [Challenges]
      security:
        - bearerAuth: []
      summary: 챌린지 상태 변경
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeUpdateRequest'
      responses:
        '200':
          description: 수정
...
  /api/v1/search:
    get:
      tags: [Search]
      summary: 통합 검색
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 검색 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    BookId:
      name: bookId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    ChallengeId:
      name: challengeId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
    RegisterRequest:
      type: object
      required: [username, password, email, nickname]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
        email:
          type: string
          format: email
        nickname:
          type: string
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/UserSummary'
    UserSummary:
      type: object
      properties:
        userId:
          type: string
        nickname:
          type: string
        role:
          type: string
        profileImage:
          type: string
...
    BookSummary:
      type: object
      properties:
        bookId:
          type: integer
        title:
          type: string
        author:
          type: string
        category:
          type: string
        avgRating:
          type: number
          format: float
    BookDetail:
      allOf:
        - $ref: '#/components/schemas/BookSummary'
        - type: object
          properties:
            description:
              type: string
            publisher:
              type: string
            isbn:
              type: string
            reviews:
              type: array
              items:
                $ref: '#/components/schemas/Review'
            wishlist:
              type: boolean
    BookCreateRequest:
      type: object
      required: [title, author]
      properties:
        title:
          type: string
        author:
          type: string
        publisher:
          type: string
        isbn:
          type: string
        category:
          type: string
        description:
          type: string
    BookUpdateRequest:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        category:
          type: string
        description:
          type: string
        isPopular:
          type: boolean
    Review:
      type: object
      properties:
        reviewId:
          type: integer
        userId:
          type: string
        nickname:
          type: string
        rating:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
...
    ChallengeSummary:
      type: object
      properties:
        challengeId:
          type: integer
        title:
          type: string
        targetBooks:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
    ChallengeDetail:
      allOf:
        - $ref: '#/components/schemas/ChallengeSummary'
        - type: object
          properties:
            description:
              type: string
            participants:
              type: array
              items:
                $ref: '#/components/schemas/ChallengeParticipant'
    ChallengeParticipant:
      type: object
      properties:
        userId:
          type: string
        nickname:
          type: string
        progress:
          type: integer
        isCompleted:
          type: boolean
    ChallengeCreateRequest:
      type: object
      required: [title, targetBooks, startDate, endDate]
      properties:
        title:
          type: string
        description:
          type: string
        targetBooks:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    ChallengeUpdateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [SCHEDULED, ACTIVE, COMPLETED, CANCELLED]
    PostSummary:
      type: object
      properties:
        postId:
          type: integer
        postType:
          type: string
          enum: [QUIZ, POLL, DISCUSSION]
        title:
          type: string
        authorNickname:
          type: string
        createdAt:
          type: string
          format: date-time
...
    SearchResponse:
      type: object
      properties:
        keyword:
          type: string
        books:
          type: array
          items:
            $ref: '#/components/schemas/BookSummary'
        posts:
          type: array
          items:
            $ref: '#/components/schemas/PostSummary'
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserSummary'
...